Refactor so that the UPRN hiding functionality is in the UPRN helpers file
